refactor(normalizer): move defaultContext
