+++ eo-phi-normalizer/eo-phi-normalizer.cabal
@@ -23,6 +23,7 @@ extra-source-files:
     report/main.js
     report/styles.css
     data/0.36.0/0.36.0.md
+    data/0.36.0/all-atoms.md
     data/0.36.0/org/eolang/as-phi.phi
     data/0.36.0/org/eolang/bool.phi
     data/0.36.0/org/eolang/bytes.phi
@@ -53,6 +54,7 @@ extra-source-files:
     data/0.36.0/org/eolang/tuple.phi
     data/0.36.0/org/eolang/while.phi
     data/0.37.0/0.37.0.md
+    data/0.37.0/all-atoms.md
     data/0.37.0/org/eolang/as-phi.phi
     data/0.37.0/org/eolang/bytes.phi
     data/0.37.0/org/eolang/cage.phi
@@ -78,6 +80,7 @@ extra-source-files:
     data/0.37.0/org/eolang/tuple.phi
     data/0.37.0/org/eolang/while.phi
     data/0.38.0/0.38.0.md
+    data/0.38.0/all-atoms.md
     data/0.38.0/org/eolang/as-phi.phi
     data/0.38.0/org/eolang/bytes.phi
     data/0.38.0/org/eolang/cage.phi
@@ -102,6 +105,7 @@ extra-source-files:
     data/0.38.0/org/eolang/try.phi
     data/0.38.0/org/eolang/tuple.phi
     data/0.38.0/org/eolang/while.phi
+    data/0.38.4/all-atoms.md
     data/0.38.4/org/eolang/as-phi.phi
     data/0.38.4/org/eolang/bytes.phi
     data/0.38.4/org/eolang/cage.phi
@@ -155,6 +159,7 @@ library
       Language.EO.Phi.Rules.Common
       Language.EO.Phi.Rules.Fast
       Language.EO.Phi.Rules.PhiPaper
+      Language.EO.Phi.Rules.RunYegor
       Language.EO.Phi.Rules.Yaml
       Language.EO.Phi.Syntax
       Language.EO.Phi.Syntax.Abs
@@ -260,6 +265,7 @@ test-suite doctests
       Language.EO.Phi.Rules.Common
       Language.EO.Phi.Rules.Fast
       Language.EO.Phi.Rules.PhiPaper
+      Language.EO.Phi.Rules.RunYegor
       Language.EO.Phi.Rules.Yaml
       Language.EO.Phi.Syntax
       Language.EO.Phi.Syntax.Abs
