+++ eo-phi-normalizer/src/Language/EO/Phi/Rules/Fast.hs
@@ -6,24 +6,10 @@ module Language.EO.Phi.Rules.Fast where
 import Data.List.NonEmpty qualified as NonEmpty
 import Language.EO.Phi.Rules.Common
 import Language.EO.Phi.Rules.Yaml qualified as Yaml
-import Language.EO.Phi.Syntax (printTree)
 import Language.EO.Phi.Syntax.Abs
-import System.IO.Unsafe (unsafePerformIO)
 
 -- $setup
 -- >>> :set -XOverloadedStrings
-
-runWithYegorRules :: (Context -> Object -> Object) -> Object -> IO ()
-runWithYegorRules f obj = putStrLn (printTree (f (defaultContext yegorRules obj) obj))
-
-yegorRuleSet :: Yaml.RuleSet
-{-# NOINLINE yegorRuleSet #-}
-yegorRuleSet =
-  unsafePerformIO $
-    Yaml.parseRuleSetFromFile "eo-phi-normalizer/test/eo/phi/rules/yegor.yaml"
-
-yegorRules :: [NamedRule]
-yegorRules = map Yaml.convertRuleNamed (Yaml.rules yegorRuleSet)
 
 withBinding :: (Context -> Object -> Object) -> Context -> Binding -> Binding
 withBinding f ctx = \case
