+++ eo-phi-normalizer/src/Language/EO/Phi/Rules/RunYegor.hs
@@ -1,0 +1,20 @@
+module Language.EO.Phi.Rules.RunYegor where
+
+import Language.EO.Phi.Rules.Common
+import Language.EO.Phi.Syntax.Abs
+import Language.EO.Phi.Rules.Yaml qualified as Yaml
+import System.IO.Unsafe (unsafePerformIO)
+import Language.EO.Phi.Syntax (printTree)
+import Language.EO.Phi.Dataize
+
+runWithYegorRules :: (Context -> Object -> Object) -> Object -> IO ()
+runWithYegorRules f obj = putStrLn (printTree (f (defaultContext yegorRules obj) obj))
+
+yegorRuleSet :: Yaml.RuleSet
+{-# NOINLINE yegorRuleSet #-}
+yegorRuleSet =
+  unsafePerformIO $
+    Yaml.parseRuleSetFromFile "eo-phi-normalizer/test/eo/phi/rules/yegor.yaml"
+
+yegorRules :: [NamedRule]
+yegorRules = map Yaml.convertRuleNamed (Yaml.rules yegorRuleSet)
