+++ eo-phi-normalizer/test/Language/EO/Phi/DataizeSpec.hs
@@ -9,9 +9,9 @@ import Test.Hspec
 
 import Language.EO.Phi (printTree)
 import Language.EO.Phi qualified as Phi
-import Language.EO.Phi.Dataize (dataizeRecursively)
+import Language.EO.Phi.Dataize (dataizeRecursively, defaultContext)
 import Language.EO.Phi.Dependencies (deepMergePrograms)
-import Language.EO.Phi.Rules.Common (defaultContext, equalObject)
+import Language.EO.Phi.Rules.Common (equalObject)
 import Language.EO.Phi.Rules.Yaml (convertRuleNamed, parseRuleSetFromFile, rules)
 import Test.EO.Phi (DataizationResult (Bytes, Object), DataizeTest (..), DataizeTestGroup (..), dataizationTests)
