+++ eo-phi-normalizer/test/Language/EO/YamlSpec.hs
@@ -5,7 +5,8 @@
 module Language.EO.YamlSpec where
 
 import Control.Monad (forM_)
-import Language.EO.Phi.Rules.Common (applyOneRule, defaultContext, equalObject)
+import Language.EO.Phi.Dataize (defaultContext)
+import Language.EO.Phi.Rules.Common (applyOneRule, equalObject)
 import Language.EO.Phi.Rules.Yaml (Rule (..), RuleSet (..), RuleTest (..), convertRuleNamed)
 import Test.EO.Yaml
 import Test.Hspec
